signed/unsigned fix

+++ source/adios2/engine/bp3/BP3Writer.cpp
@@ -118,7 +118,8 @@ void BP3Writer::Flush(const int transportIndex)
 void BP3Writer::Init()
 {
     InitParameters();
-    if (m_BP3Serializer.m_Parameters.NumAggregators < m_BP3Serializer.m_SizeMPI)
+    if (m_BP3Serializer.m_Parameters.NumAggregators <
+        static_cast<unsigned int>(m_BP3Serializer.m_SizeMPI))
     {
         m_BP3Serializer.m_Aggregator.Init(
             m_BP3Serializer.m_Parameters.NumAggregators, m_Comm);
