+++ source/adios2/engine/bp4/BP4Writer.cpp
@@ -123,7 +123,8 @@ void BP4Writer::Flush(const int transportIndex)
 void BP4Writer::Init()
 {
     InitParameters();
-    if (m_BP4Serializer.m_Parameters.NumAggregators < m_BP4Serializer.m_SizeMPI)
+    if (m_BP4Serializer.m_Parameters.NumAggregators <
+        static_cast<unsigned int>(m_BP4Serializer.m_SizeMPI))
     {
         m_BP4Serializer.m_Aggregator.Init(
             m_BP4Serializer.m_Parameters.NumAggregators, m_Comm);
